Norminette for glmath/srcs/primitives
